Extract KomadaiArtist class to its own file
